Adds single step action
This makes the function first-class, and runsa s ingle-step
function/reducer. THe tricky thing here is with inheritance -- we have a
`run` and an `update` but we don't call it as it subclasses `Action`. We
will be cleaning this up shortly (insomuch as we can) and thinking
through the abstraction soon.

This is a solution for #11