Uber commit to make rebasing simpler:
Adds serialization hook for lcel example (+12 squashed commits)
Squashed commits:
[966cffd] Fixes python repl response

If this doesn't work I should
revert to the lecl response structure of their python repl
function.
[ac0a0a1] Gets it almost working

the LCEL one uses the deprecated function paths
of open ai API. I'm trying to use the new tool one here.
It's a bit inconsistent it seems -- I did manage to get
it to draw a graph, but it failed to finish because I hadn't
set up how it should say to finish things.

Otherwise a UI to debug this would be great.
[12feac6] Adds working lcel multi-agent example

Will now take this and rewrite with Hamilton/more obvious
constructs.
[4ef64f2] Adds beginnings of multi-agent example

Need to figure out the application state model and
what to update, i.e. wire through messages history
properly.

Otherwise rather than having a specific tool node, like
the langgraph example, I added that call to the hamilton
DAG for it -- could always change it to mirror their stuff.
[f3ffe95] Adds telephone example README
[03b6d07] Addresses comments
[11d0b0f] Fixes formatting
[17ffa67] Adds telephone app example

This uses the dataflows from the hub and then shows two ways
to integrate Hamilton. One via the Hamilton integration, the other
via the function API.
[557b24b] Adds image telephone example

This uses the dataflows defined on the hub to create the image
telephone app.
[0233a95] Fixes cowsay application __main__

the variables were round the wrong way.
[dbeddb0] WIP
[0ba0158] WIP (+1 squashed commit)
Squashed commits:
[bc53419] Initializes telemetry UI

Design:

1. CRA + tailwind (they don't like CRA but I really don't care)
2. prettier + eslint
3. Add to pre-commit-config.yaml rahte rthan husky (which is very
   painful IMO)