Adds backend tracking server to Burr
This is a little quick/dirty. The intention is that the local data model
will work and remain stable through releases (if not, we will have
automated migration). The contract between server and UI is not stable
and will likely change. This is not included in the semantic versioning
yet for those reasons.

Basic design:
1. FastAPI for server
2. Pydantic for data types
3. Aiofiles + quick file-list for db (filesystem) access
4. Simple BE abstraction to start forcing the difference between API and
   implementation (though this will change)
5. uvicorn to run