Solves #31 -- enables deletion operation
We weren't respecting it before. This handles that. This just measures
the delete operations and reapplies it. That said, this is not the best
way of doing things -- see #33 for a more involved approach.

Note we've also relaxed the restrictions on modifications for
delete/write -- this is necessary for the current workaround.