Adds function to get state to tracker
So that we can pull the last state given an ID from the UI.

Updates examples and adds a new one that mirrors the
LCEL example with the tool calling. This variant works.
Where as the Hamilton DAG also doing tool calling somehow doesn't quite work
as well. Still digging in there. Need a run diff tool now... or a way to compare runs...