Added load to return typeddict and saving an error
So the return of load was getting out of hand. Added
TypedDict response. I feel like that's a bit easier to extend.

Also decided that save should have a status field. So
the persister can make a choice whether to save it or not,
and then likewise when loading someone can determine
what behavior they want.

Otherwise possible edge cases / bugs to iron out:

 - async/streaming -- not tested
 - sequence id needs to be done prior to the step
 - also reloading from an error nuking  `__prior_step` means
that it isn't set until the third iteration weirdly. Haven't had a chance
to dig.