Adds additional safeguards around writing state
If your action declares a state write that it does not apply, it will
now error out. This is backwards compatible, as that was always
considered undefined behavior, and bad practice.

Note that this does not mess with deleting state in actions, as that is
not declared in the writes() field. We will want to handle it later.