Adds command to create a test case
The idea is that someone is in the tracker and wants to fix/create
a test around some part of the application.

They can then use the command line to bootstrap it. We could
do this in the UI too eventually.

We assume they want to then have a JSON file with input state
and expected output state to validate against. I think this
is a nice way to do it. Then it's all checked in and they can run
it in CI easily...

This command assumes a local tracker for now. In the future
we can add a yaml/config file that can instantiate a persister
to load data instead.

Otherwise we assume pytest, we could create a unittest stub too,
and also that users will like the file drive approach.