Updates telephone example
To make it a top-level example.

Simplifies it a little and splits out the two implementations
to make the code simpler to look at.

Adds integrations to be skipped for validation for now.
Proper fix would be to mark directories as example ones
and only validate those...

Adds an example hook to save images locally.