Adds modular SERDE
This code introduces a core module for SERDE.
State relies on this to know how to serialize and deserialize itself.

A design decision made was that we will use singledispatch style
registration based on types for serialization, and a string key value
for deserialization. Serialization always outputs a dictionary that
can then be JSON serialized somewhere. A key is added to that
to know how to deserialize things.

Field level SERDE specification is still possible, but hasn’t been
implemented yet.

Otherwise people interacting with State object should use the methods
defined on it. The core.serde module can be used in other
contexts easily too — so if trackers and persisters want to utilize it they can.

One constraint is that all SERDE functions take in **kwargs. That way people
could wire through some interesting args to do more interesting serialization,
e.g. like for pickle, or pandas. The kwargs are really a dict of kwargs, where
each serde function/module should namespace its kwargs to avoid clashes.

If in the future we need to change what is registered as SERDE, that’s also possible
to adjust. The only limitations this design imposes is a global way to do SERDE.
Field level will come, and make it even more customizeable, otherwise we think
this approach should cover most use cases.

This change also fixed a bug in the tracker when looking for sequence ids, because
it hadn’t anticipated spans…

Squashed commits:
[6cbc9ee] Flips print to debug
[b9b7dda] Fixes broken tests due to deps
[9daf325] Updates code and adds tests

This adds tests and removes code.
[20579ac] Fixes bugs with incomplete file log state

File logs could be incomplete. So this checks to find
the last entry that is valid for that particular context.
[47d3288] WIP: Adds more docs and adds pickle

This commit will be cleaned up.

But adds more docs, and a way to register
classes to be pickled!
[1a1efb2] Refactors serde to proper place and adds POC

This is WIP -- but it seems to work via unit tests.

Uses **kwargs for future proofing, but also allowing
one to pass in values to serde -- idea is that each
would have some namespace of kwargs.

Just now need to test it for the tracker.

TODO: allow one to register field level state serde.
[3762aa9] Adds prototype module to handle serde

Singledispatch for type serialization. For non-primitive types, make them return a dict with a serde key.

Then use a string value based mechanism to deserialize. This way we can be "field agnostic" during deserialization.

Otherwise deserialization without knowing what it is, is a challenge.
Thus the idea to include a key in a dict for it.

Otherwise the only other way for deserialization is to field
registries...

TODO:
- determine module naming and location
- actually wire in
- proper tests
- how to override these
- think about any other edge/error cases