Adds validation for conditions
Adding variables that were not in state to the condition results in
unclear error message (especially `expr`. This fixes it by validating
beforehand to give a clearer message.

Note that if there's a smart "or" statement, it might error out. I think
this is less than ideal, but we can see if anyone complains.