Wires through application context to nodes that request it
A node can now query for application context by setting an input
as __context -- this is of type ApplicationContext. This will allow
forking (+ any other information we may need from inside the node).
This is one of two __ variables (the other being tracer).

The dependency factory is a bit in flux but the way these are
constructed is not yet public-facing (we cannot arbitrarily add global
injectors), so we can change it.

This also fixes a bug in which dependency factories were not allowed to
be optional/have defaults. This enables the user to build out a
function that can be run outside the app (with a null context)..