Allows application to wire through spawning parent
This enables an application to specify the app that spawned it.

- Adds tracking of spawning parent to child app. This enables us to track from child app the parents that spawned it
- Adds pointers from parent app to child app. This does a little trickiness with the local tracker, but will allow the
UI to display bidirectional links for forks/spawns.
- Adds visualization for spawning parents to App list. This allows us to expand/focus on children in the app list. This also hides the partition key if it doesn't exist

See https://github.com/DAGWorks-Inc/burr/issues/211

Adds bidirectional linking to the UI for spawning/forking apps

This adds the proper UI component + fixes logging/server to log
correctly when spawning.