Fixes for test case creation
1. Made partition key optional.
2. This was not updated to use the latest SERDE methods.
3. It was also missing the ability to plugin custom SERDE
functions.

For (3) we enable one to pass in a python file, that will
then be imported, or a fully qualified module name to
be imported.

We add an example to script that runs the custom serde
application whose output can then be used to test
passing in the python file for creating a test case
from that run.