Adds default reads/writes to burr actions
This allows you to specify defaults if your action does not write. In
the majority of cases they will be none, but this allows simple (static)
arbitrary values. This specifically helps with the branching case --
e.g. where you have two options, and want to null out anything it
doesn't write. For instance, an error and a result -- you'll only ever
produce one or the other.

This works both in the function and class-based approaches -- in the
function-based it is part of the two decorators
(@action/@streaming_action). In the class-based it is part of the class,
overriding the default_reads and default_writes property function

We add a bunch of new tests for default (as the code to handle multiple
action types is fairly dispersed, for now), and also make the naming of
the other tests/content more consistent.

Note that this does not currently work with settings defaults to
append/increment operations -- it will produce strange behavior.
This is documented in all appropriate signatures.

This also does not work (or even make sense) in the case that the
function writes a default that it also reads. In that case, it will
clobber the current value with the write value. To avoid this,
we just error out if that is the case beforehand.