Adds hooks for application interface calls
See https://github.com/DAGWorks-Inc/burr/issues/203.

We allow pre/post execution for the following:

- step/astep
- iterate/aiterate
- run/arun
- stream_result/astream_result

These enable us to add tracing in with these functions.

We use a context manager to do the first three -- the last one is funky,
as we call after the execution is complete. We use the pre/post call
functions from the afforementioned context manager to set it up, and
pass it to the callbacks that get executed at the end of streaming. This
way the trace shows up as the full stream.