Adds OpenTelemetry integration.
Two approaches:
1. Log Burr -> Otel -- simple adapter
2. Log any Otel traces inside a burr step to the Burr UI

This is overall fairly sound, although there is a bit of global state
that might be dealt with.

Design decisions:

1. Use a custom span processor
2. Store state of a map of span_id -> step
3. Propogate that when processing spans

It's a bit odd as the span processor should just be writing/queuing, but
the way I'm thinking of this is that the writing is converting OTel
spans to Burr spans, so we have to do that somewhere. Why not when
processing? Then we can just cascade down.

Any spans outside of a Burr step are ignored, as we are not a full OTel
provider.