Fixes deployment target on Burr to be more modular
We remove the CLI stuff, and enable the user to install it separately.
The tracking-server-s3 doesn't need it, so we have it in a different
target. This also means they don't have to install burr[start] to run
the server, which is preferable (as versions of boto3 and aiobotocore
conflict)