Fixes finally block in streaming
We hit a nasty context var error -- it said that the context var was set
in a different context. In reality this happens inside the finally
block, which is run at the very end, after the context var. So, instead
of resetting to the token, we reset to None, which has a similar effect
(and is after the whole thing is done, so it doesn't really matter).

Note this coud potentially have interesting implications for Burr within
BUrr (non threaded), but we're going to want to htihnk that through
later.