Adjust test case command
This now correctly handles cases where there was no failure in the action.

When an action fails we save the initial state that went in. However, when
an action succeeds, we would only store the final state for that action.
So now, if the action was successful, we go back to the prior sequence
ID to get its state as the input, and use the state from the action as the
desired final state. This assumes you want to say "hey here's expected
behavior, I want to enshrine it".

Adjusts docs with new option.