Adds schema tracking to action
This is just a Quick wrapper class to represent a schema. Note that this is currently used internally,
just to store the appropriate information. This does not validate or do conversion, currently that
is done within the pydantic model state typing system (which is also internal in its implementation).

We will likely centralize that logic at some point when we get more -- it would look something like this:
1. Action is passed an ActionSchema
2. Action is parameterized on the ActionSchema types
3. Action takes state, validates the type and converts to StateInputType
4. Action runs, returns intermediate result + state
5. Action validates intermediate result type (or converts to dict? Probably just keeps it
6. Action converts StateOutputType to State

Also we don't have this split out into two classes (input/output schema), but we will likely do that shortly.