Wires through persistence to spawned children apps
This enables us to use persisters from the parent in spawned apps.

It roughly works as follows:
1. We add a copy() method to the persisters
2. We pass hte persisters into the context input to the child
3. The child then clones the persisters/trackers
4. The child has a setting that allows for either cascading
or overwriting persisters/trackers/whatnot (cascade is the default).
5. We use the same instance of persister/loader *if* the parent is using
   the same instance

Note that this also improves some asynchronous generator stuff -- it's a
bit complicated, but it allows you to define asynchronous *or*
synchronous generators for the task-producing subclasses of parallel
actions. The type is detected at runtime, allowing flexible
implementation.