Refactors Redis Persister
This adds a new class RedisBasePersister that the old one inherits from.

The reason to do this is that if someone wants to test things, then being
able to inject a mock redis client makes this simpler. So refactoring
to enable that while still being backwards compatible.

Appropriately marked things as deprecated and tests have been updated.