Adds more thorough pytest example (#442)
Adds example on using pytest for testing - explaining pytest,
but also doing a more realistic example.

This example just shows how to use the tools so to speak. It doesn't 
talk about what should be evaluated much at all. But hopefully it
should give people an idea of what to do and how...

This example should be updated as we do moe.

Adds SDLC images and verbiage too.