Adds action tagging
This enables us to use tags as aliases for actions. We may add future
tags (E.G. __requires_inputs), but for now this is all done on behalf of
the user. The idea is to have a very simple polymorphic interface -- we
can (for example) say all nodes tagged with "text_return" fullfill the
same role (displaying markdown for the user).

See #468 for more details.