Deep researcher AI reasoners are increasingly popular commercial products, with google and Open AI both offering one.  This fact suggests the possibility of open source implementations.  This example's particular research flow was adapted from the langchain Local Deep Researcher (https://github.com/langchain-ai/local-deep-researcher), replacing langgraph with Burr as the graph framework of choice. In the Burr implementation, calls to a local model are also replaced by api calls, for the convenience of those who don't currently have the hardware to host a local model.
The patterns in the code are basic Burr graph patterns, as described in
the online documentation (https://burr.dagworks.io/). The UI is written in React
and based on the example email assistant UI and the example chatbot UI. Here I
consulted the developer guide (https://burr.dagworks.io/contributing/) quite a bit.
The React wrapper for the python API can be regenerated using
the script telemetry/ui/scripts/client-gen.sh

Future developments of the workflow and the UI could include adding
concurrency and streaming to the model, both to increase its power and
to improve the user experience.